requirements.txt
# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0

# Database - Updated for Python 3.13 compatibility
sqlalchemy==2.0.23
psycopg[binary]==3.2.3
pgvector==0.3.6
alembic==1.12.1

# OAuth and HTTP
httpx==0.25.0
oauthlib==3.2.2

# AI and ML
anthropic==0.39.0
torch==2.1.0
torchvision==0.16.0
open-clip-torch==2.24.0
numpy==1.24.0
# Note: Install PyTorch and CLIP separately based on your system:
# pip install torch torchvision open-clip-torch --index-url https://download.pytorch.org/whl/cpu

# Image Processing
pillow==11.0.0

# Configuration and Utilities
python-dotenv==1.0.0
python-multipart==0.0.6

# Development
pydantic==2.5.0
typing-extensions==4.8.0

# Testing
pytest==7.4.0
pytest-asyncio==0.21.1
pytest-cov==4.1.0
pytest-mock==3.11.1
faker==19.6.0